Move async function arguments into spawned futures
Fixes #3042

Previously, any exported async functions which borrowed their arguments would fail to compile, because the values being referenced are stored on the stack of the initial synchronous call and then dropped, leaving nothing for the spawned future to reference.

This moves the referenced values into the spawned future, fixing that.