feat: Implement `JsValue.downcast_unchecked`.
Since the introduction of the new `JsValue.as_ptr` method, it is now
possible to implement a new handy method: `JsValue.downcast_unchecked`
(which is unsafe by nature because no check is applied).

`downcast_unchecked` will downcast the `JsValue` to `T::Anchor` where
it implements `RefFromWasmAbi<Abi = u32>`. It returns a `Ref<T>` which
can be deref to `&T`.