Call __destroy_into_raw everywhere a value is consumed
When using weak references the object needs to be unregistered from the
finalization registry when we pass it to a Rust method that takes in the
object by value. This wasn't happening, which meant that the finalizer
was called despite the object already being freed. For me this
manifested as a 'recursive use of an object detected which would lead to
unsafe aliasing in rust' exception.