Add bindings to `maplike` interfaces
`lazy_static` is used to create some static extended attribute in order
to add it to the signature of the generated operation during the first
pass. Without this, the attribute cannot live long enough, as it must be
stored in a `Vec`.

I tried to use `Cow` to allow the attribute to be owned by `Signature`,
but that makes the borrow checker unhappy somewhere in `lib.rs` in a
seemingly unrelated place. This is very curious but I figured this would
be the easiest way to get it working for now.