Apply `#[allow(dead_code)]` to `#[wasm_bindgen]`'d functions
Fixes #2726

`#[wasm_bindgen]`'d functions are always used, since they're exported to JS.
However, that doesn't happen when not building for the wasm target, and so a
dead code warning is shown if the function isn't used for anything else. This
fixes that by adding `#[allow(dead_code)]` to `#[wasm_bindgen]`'d functions to
supress the warning.