Apply `#[allow(dead_code)]` to `#[wasm_bindgen]`'d functions (#3180)
* Apply `#[allow(dead_code)]` to `#[wasm_bindgen]`'d functions

Fixes #2726

`#[wasm_bindgen]`'d functions are always used, since they're exported to JS.
However, that doesn't happen when not building for the wasm target, and so a
dead code warning is shown if the function isn't used for anything else. This
fixes that by adding `#[allow(dead_code)]` to `#[wasm_bindgen]`'d functions to
supress the warning.

* fix syntax error on rust 1.56

I'm not sure what exactly was invalid about it, but `quote! { ... }.to_tokens(...)` was causing syntax errors on Rust 1.56 in CI for some reason, so I changed it to `tokens.extend(...)`, which is used elsewhere in the file and so should work.