Exclude `#[wasm_bindgen]`'d functions in `discover_main`
This is a partial fix for #3199. I couldn't find any way to reliably detect whether a `main` function was actually a proper Rust `main` function, but the common case of a `#[wasm_bindgen]`'d function named main happening to have the same signature can be avoided, by just checking whether the function is a `#[wasm_bindgen]` export before marking it as the start function. This PR does that.

It's still possible for a function to be mistakenly treated as `main` if it's exported manually, but that's much less likely to occur.

Resolves #3199.