Exclude `#[wasm_bindgen]`'d functions in `discover_main` (#3206)
* Exclude `#[wasm_bindgen]`'d functions in `discover_main`

This is a partial fix for #3199. I couldn't find any way to reliably detect whether a `main` function was actually a proper Rust `main` function, but the common case of a `#[wasm_bindgen]`'d function named main happening to have the same signature can be avoided, by just checking whether the function is a `#[wasm_bindgen]` export before marking it as the start function. This PR does that.

It's still possible for a function to be mistakenly treated as `main` if it's exported manually, but that's much less likely to occur.

Resolves #3199.

* Use `self.adapters.exports`, not `self.function_exports`

`self.function_exports` is the list of _all_ exported functions, not just the `#[wasm_bindgen]`'d ones. So, this was completely broken before.

I'm still not 100% sure if `self.adapters.exports` is the right thing to be using though.