Improve error message when `#[wasm_bindgen(getter_with_clone)]` is used on non-`Clone` JS types (#3295)
* Improve error message when `#[wasm_bindgen(getter_with_clone)]` is used on non-`Clone` JS types

Fixes #3291

Previously, the implementation of that attribute would call `.clone()` on the field to try and clone it. The problem is that imported JS types implement `Deref<Target = JsValue>`, which meant that calling `.clone()` would auto-deref the type into a `JsValue`, and then clone that. So, instead of a '`Foo` does not implement `Clone`' error, you would get a cryptic error about a type mismatch.

This fixes that by clarifying the `clone` call to use the `<Foo as Clone>::clone(&val)` syntax instead. I also added some spans so that the error message will point out the field that's causing the problem.

* fmt