Update expected results for UI tests (#3312)
* Update expected results for UI tests

Apparently, #3293 didn't quite update them properly, causing CI to fail. I'm not sure why CI passed when checking the PR but failed on `main`, but this should fix the CI failure.

* Use updated trybuild

It looks like trybuild moved where it puts its output, and so CI's up-to-date trybuild and my older local trybuild disagreed on what a path should be.