Don't rely on the name section for detecting `main`
Fixes #3362

In some cases, the name section can fail to be parsed because one of the names is too long, which then causes the main function to not be detected because we don't know what any functions' names are.

However, `main` is also exported with the name `main`, so we can look for an export named `main` instead to avoid relying on the name section.