Don't install `wasm-opt` in CI (#3366)
* Don't install binaryen in CI

This is a workaround for rustwasm/wasm-pack#1247, which causes `wasm-pack` to fail to run if a local version of `wasm-opt` is installed. That in turn caused the `build_examples` CI job to break.

This PR fixes that by not having it download binaryen, instead letting `wasm-pack` install it itself, which still works.

* Put back the binaryen-installing step, but only install wasm2js instead of the whole thing.

I also updated to binaryen 112 while I was at it.