Fix race condition in reference tests (#3373)
The reference test runner was attempting to compile all of them in parallel, with all of their outputs being written to `target/wasm32-unknown-unknown/debug/reference_test.wasm`.
That led to a race condition if this happened:

* test a finishes compiling, emits `reference_test.wasm`
* test b finishes compiling, emits `reference_test.wasm`
* test a runs `wasm-bindgen` on the `reference_test.wasm` that test b just emitted.

This led to the results of reference tests getting mixed up.

For whatever reason, this only happens on my Mac laptop and not my Linux desktop. It could be an OS difference, or just because the laptop's a lot faster, I don't know.

I solved it by just not running them in parallel, since parallelising them didn't do much anyway thanks to `cargo` locking the build directory.