Add `sorted_iter_mut` for NonstandardWitSection.adapters for determinism
In the `externref` code, we call `export_xform` while iterating
over `adapters` with `iter_mut`. Calling `export_xform` can result
in shims, with the id depending on the order in which we iterate
over the exports.

To make the final .wasm output deterministic, I've changed this to
use a new `sorted_iter_mut` method (based on `sorted_iter`).